+++ src/App.css
@@ -2,4 +2,10 @@
   font-family: Inter, system-ui, Avenir, Helvetica, Arial, sans-serif;
   line-height: 1.5;
   font-weight: 400;
+
+  --light-primary: #ececf1;
+}
+
+body {
+  background-color: var(--light-primary);
 }
